+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- OpenTelemetry tracing support via `NewClientWithOTEL` and `NewClientWithResponsesWithOTEL`
+
 ## [1.1.0] - 2026-01-14
 
 ### Changed